+++ NEWS.md
@@ -9,6 +9,7 @@ NEWS
 - Added `smd_calc` and `ses_calc` for just calculating the standardized effect sizes (no tests).
 - Default CIs for for SMDs are now NCT rather than the Goulet method.
 - `compare_smd` can be supplied with user provided standard errors.
+- Add `log_TOST` function for comparing ratios of means.
 
 # TOSTER v0.5.1